make sure fs storage adapter returns false on a copy fail

+++ src/seer/automation/codebase/storage_adapters.py
@@ -81,7 +81,12 @@ class FilesystemStorageAdapter(StorageAdapter):
         workspace_path = self.get_workspace_location(self.repo_id, self.namespace_id)
         storage_path = self.get_storage_location(self.repo_id, self.namespace_slug)
 
-        shutil.copytree(storage_path, workspace_path, dirs_exist_ok=True)
+        if os.path.exists(storage_path):
+            try:
+                shutil.copytree(storage_path, workspace_path, dirs_exist_ok=True)
+            except Exception as e:
+                autofix_logger.exception(e)
+                return False
 
         return True
 
@@ -123,27 +128,31 @@ class GcsStorageAdapter(StorageAdapter):
         workspace_path = self.get_workspace_location(self.repo_id, self.namespace_id)
         storage_prefix = self.get_storage_prefix(self.repo_id, self.namespace_slug)
 
-        blobs = self.get_bucket().list_blobs(prefix=storage_prefix)
-        blobs_list: list[storage.Blob] = list(blobs)
-        for blob in blobs_list:
-            if blob.name:
-                filename = blob.name.replace(storage_prefix + "/", "")
-                download_path = os.path.join(workspace_path, filename)
+        try:
+            blobs = self.get_bucket().list_blobs(prefix=storage_prefix)
+            blobs_list: list[storage.Blob] = list(blobs)
+            for blob in blobs_list:
+                if blob.name:
+                    filename = blob.name.replace(storage_prefix + "/", "")
+                    download_path = os.path.join(workspace_path, filename)
 
-                if not os.path.exists(os.path.dirname(download_path)):
-                    os.makedirs(os.path.dirname(download_path))
+                    if not os.path.exists(os.path.dirname(download_path)):
+                        os.makedirs(os.path.dirname(download_path))
 
-                blob.download_to_filename(download_path)
+                    blob.download_to_filename(download_path)
 
-                # Update the custom time of the blob to the current time
-                # We use custom time to track when the file was last used
-                # This is to manage the lifecycle of the files in the storage
-                blob.custom_time = datetime.datetime.now()
-                blob.patch()
+                    # Update the custom time of the blob to the current time
+                    # We use custom time to track when the file was last used
+                    # This is to manage the lifecycle of the files in the storage
+                    blob.custom_time = datetime.datetime.now()
+                    blob.patch()
 
-        autofix_logger.debug(
-            f"Downloaded files from {storage_prefix} to workspace: {workspace_path}"
-        )
+            autofix_logger.debug(
+                f"Downloaded files from {storage_prefix} to workspace: {workspace_path}"
+            )
+        except Exception as e:
+            autofix_logger.exception(e)
+            return False
 
         return True
 
@@ -151,27 +160,33 @@ class GcsStorageAdapter(StorageAdapter):
         workspace_path = self.get_workspace_location(self.repo_id, self.namespace_id)
         storage_prefix = self.get_storage_prefix(self.repo_id, self.namespace_slug)
 
-        blobs = self.get_bucket().list_blobs(prefix=storage_prefix)
-        for blob in blobs:
-            # Delete the existing blobs in the storage prefix
-            blob.delete()
+        try:
+            blobs = self.get_bucket().list_blobs(prefix=storage_prefix)
+            for blob in blobs:
+                # Delete the existing blobs in the storage prefix
+                blob.delete()
 
-        for root, dirs, files in os.walk(workspace_path):
-            for file in files:
-                file_path = os.path.join(root, file)
-                relative_path = os.path.relpath(file_path, workspace_path)
-                blob_path = f"{storage_prefix}/{relative_path}"
+            for root, dirs, files in os.walk(workspace_path):
+                for file in files:
+                    file_path = os.path.join(root, file)
+                    relative_path = os.path.relpath(file_path, workspace_path)
+                    blob_path = f"{storage_prefix}/{relative_path}"
 
-                blob = self.get_bucket().blob(blob_path)
+                    blob = self.get_bucket().blob(blob_path)
 
-                # Update the custom time of the blob to the current time
-                # We use custom time to track when the file was last used
-                # This is to manage the lifecycle of the files in the storage
-                blob.custom_time = datetime.datetime.now()
+                    # Update the custom time of the blob to the current time
+                    # We use custom time to track when the file was last used
+                    # This is to manage the lifecycle of the files in the storage
+                    blob.custom_time = datetime.datetime.now()
 
-                blob.upload_from_filename(file_path)
+                    blob.upload_from_filename(file_path)
 
-        autofix_logger.debug(f"Uploaded files from workspace: {workspace_path} to {storage_prefix}")
+            autofix_logger.debug(
+                f"Uploaded files from workspace: {workspace_path} to {storage_prefix}"
+            )
+        except Exception as e:
+            autofix_logger.exception(e)
+            return False
 
         return True
